Remove test timeout from engine_testapp
engine_testapp had support for specifying a test timeout
by using -t command line option. Over the years it's
been modified and in it's current form it no longer works
as all it does is to set a variable we never check.

These days we run the automated builds under Jenkins
which monitors the timeout outside the binary so we
can just remove the code.

Change-Id: I5118d9d84b6581b7253cab14642a590be38521e6
Reviewed-on: http://review.couchbase.org/104688
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>