MB-32974 [SR]: Add consumer_name to DCP_OPEN
Update DCP_OPEN handling in memcached to check for and validate a JSON
value.

Update ep-engine to check if the JSON value contains the key
'consumer_name' (for Consumer connections), and if so use that as the
consumer_name when negotiating control messages with the producer.

Assuming ns_server support
(i.e. http://review.couchbase.org/#/c/104483/), this allows ns_server
to correctly set the toploogy and node names and for multiple replicas
to be supported for SyncWrites.

Change-Id: I82cf0af4ce61b1459837cd8f915a703fa2f6e567
Reviewed-on: http://review.couchbase.org/104705
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>