MB-24377: Get all keys return defaultC when not collections aware
Currently, get all keys will return keys from any collection when
called. In the UI Documents page, we display collections keys as
a result, even though the UI is not collections aware. This is an
issue as the keys return an error when the UI attempts to lookup
the document, and the UI fails to display anything if it detects a
duplicate key.

Fix this by only returning keys from the default collection if the
client is not collection aware.

Change-Id: I9282448b13de4a04bd84657694d680b0b7d226d2
Reviewed-on: http://review.couchbase.org/104747
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>