MB-32784: 3/5 Switch to new meta-data for dropping collections
Note this patch maintains two concurrent collection's meta-data
mechanisms and partially changing over to use the new. This patch
deliberately does not delete all of the code which is managing the old.
Complete removal of the old mechanism will happen in a later patch to
keep the diff size more manageable.

This patch switches over to the new persisted meta-data with two major
areas affected.

The VB::Manifest now has a more stream-lined task, it is no longer
re-used in VB::ScanContext and VB::EraseContext for collection drop
and backfill 'pruning', the object now just manages the flow of CRUD
operations at the front of house. The VB::Manifest no longer has to
manage a collection state from open -> deleting -> gone, the
VB::Manifest now only stores open collections. The fallout from this
change is the removal of a number of bits of the VB::Manifest such
as completeDeletion(), nDeletingCollections and the greatestEndSeqno
"optimisation" (never proven to make a difference anyway...). The
CRUD paths now just need to lock the manifest and if the requested
collection-ID is in the map, the collection is writable.
isLogicallyDeleted still utilises seqno comparison for certain paths so
we can support collection flush.

The second major affected is KVStore, with couch-kvstore now
utilising the new meta-data for dropping a collection from storage.

Change-Id: Ib0aca6c9cded8f3ee01f2dfecec93b180adea045
Reviewed-on: http://review.couchbase.org/104784
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>