MB-32484: Send ABRT to memcached_testapp on timeout
To attempt to identify the cause of the intermittent timeouts in
memcached_testapp, use the `timeout` program to signal
memcached_testapp with SIGABRT when we timeout; to allow a core dump
to be generated of where the process is stuck.

Using a timeout (240s) before the CTest timeout of 300s below so
`timeout`'s timeout fires first.

Change-Id: Ice3ce4cd35b2ecad32e4b9419bd37cbf9c493f83
Reviewed-on: http://review.couchbase.org/104812
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>