MB-32784: 4/5 Stream-line collection system events
Note this patch contains two concurrent collection's meta-data
mechanisms and now fully changes over to use the new. This patch
deliberately does not delete all of the code which is managing the old.
Complete removal of the old mechanism will happen in a later patch to
keep the diff size more manageable.

The old meta data mechanism relied on VB::Manifest generating the entire
value that will be persisted (a snapshot of the entire manifest). This
patch changes to a more streamlined approach in that each system event
only carries the data it needs, for example a create collection event
contains only data about that single creation. This results in the
system events stored in the vbucket index to be much smaller and the
overall mechanism for creating the items and storing them to be much
simpler.

Change-Id: Ibdf926a753690c4e6ebf442a9d728583f04376c2
Reviewed-on: http://review.couchbase.org/104830
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>