MB-33038 [SR]: DCP Consumer knows if Producer supports Sync Repl
Enhance the DCP Sync Replication negotiation for letting the Consumer
know if the Producer supports Sync Repl.

This will be used at Consumer to decide if a DCP_SEQNO_ACK must be sent
over the connection. Note that, while a pre-MH Producer will never send
SyncWrites (and so, the Consumer will never need to SEQNO_ACK), during
cluster-topology changes we may need to distinguish between pre-MH and
MH Producers. E.g.:
- nodes Active and Replica, both MH
- Replica has some Prepares in memory, not yet ack'ed
- failover + NewActive: now NewActive is a pre-MH node
- at this point a NewActive->Replica connection is established, but we
  do not want to send any SEQNO_ACK to the Producer

Also, the new feature can be used as a general way for a Consumer to
distinguish between MH and pre-MH Producers, which can be useful at
Online Upgrade.

Change-Id: I42d65a4a0ca8d8805d2f69e550034f399b4d4ad6
Reviewed-on: http://review.couchbase.org/104823
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>