MB-32484: EWB_Engine: Avoid missed notify() for pending_io_ops
EWB_Engine::process_notifications() does not use a predicate in its
condvar.wait() call. This can result in missed notifications if this
function gets a spurious wakeup, as it'll just end up waiting forever.

Note: this is believed to be the cause for random memcached_testapp
hangs; as there are pending client io ops which are not notified.

Change-Id: Id49b150e8f7be55cb5bdb11ad0ff28101fe51529
Reviewed-on: http://review.couchbase.org/104894
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>