Audit uses of condition_variable.wait(lock) in tests
The single argument overload of std::condition_variable::wait (without
a predicate) is error-prone, as it can result in lost notifications if
the condition variable is spuriously woken up. See
http://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rconc-wait

Audit our uses of wait() without a predicate in test code; fixing any
instances where they are incorrect, and simplifying APIs to make it
harder to use incorrectly.

Change-Id: Ic401f2d17103603c1ea1805359477a25945897f8
Reviewed-on: http://review.couchbase.org/104897
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>