MB-33044: Log the content of the unparsable RBAC db
We've seen errors where RBAC fails to parse the provided RBAC
database due to format errors or similar. Given that we don't
know what the content is, it's hard to say why we fail.

Enhance the logging to dump the content in the logs so that
we may figure out why this happens.

Change-Id: I26f6d6f94a0378d9761338eab6090dce064a79bc
Reviewed-on: http://review.couchbase.org/104896
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>