MB-32591: Expel checkpoint items that are no longer required
This patch introduces the ability to expel items from the oldest
checkpoint that still has cursors in it.  This can include the open
checkpoint.

Items are expelled items from checkpoints to reduce memory requirements.
This is achieved by identifying the oldest checkpoint that has cursors
still in it.  Then identifying where the first cursor is located. For
example in the diagram below it is seqno 1004.

                  low                                high
           dummy seqno                              seqno
             |     |                                  |
             |     |                                  |
            \./   \./                                \./
         --------------------------------------------------
         | 1000 | 1001 | 1002 | 1003 | 1004 | 1005 | 1006 |
         --------------------------------------------------
                                       /.\           /.\
                                        |             |
                                        |             |
                                     cursor 1      cursor 2

It then expels items from the start of the checkpoint upto and including
where the first cursor is located.  In the diagram this means expelling
1000, 1001, 1002, 1003 and 1004. A new dummy is created at the position
of where the last cursor pointed and is assigned the same seqno as the
previous dummy.

After the expel operation the checkpoint in memory is follows.

           new   low     high
          dummy seqno   seqno
            |      |      |
            |      |      |
           \./    \./    \./
         ---------------------
          1000 | 1005 | 1006 |
         ---------------------
           /.\           /.\
            |             |
            |             |
         cursor 1      cursor 2

The number of items in the checkpoint remains unchanged after expelling
i.e. in the above example the full checkpoint remains:

        -------------------------------------------
        | 1001 | 1002 | 1003 | 1004 | 1005 | 1006 |
        -------------------------------------------
                                /.\           /.\
                                 |             |
                                 |             |
                              cursor 1      cursor 2

If a checkpoint contains a single mutation then it is not expelled.
Also if the cursor is pointing to a meta-item the position to expel
from is moved backwards until either a mutation item or the dummy item
is reached.

The ability to expel items is controlled by a configuration parameter
"chk_expel_enabled".  The parameter is defaulted to true, however is
dynamic so can be disabled at runtime if required.

Change-Id: I846aa70541d1f6f3952cd53334efe6698d683577
Reviewed-on: http://review.couchbase.org/104945
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>