Let EWB use the EngineMap from the core
This removes all of the problems related to when to unload the
shared object.

Previously ewouldblock engine didn't try to unload the engine
as part of bucket deletion (instead it just freed the bucket
handle). The next load of the bucket would be a "noop" as the
shared object was already loaded. As part of moving to
unique_ptr's for the shared objects we would delete and
close the dll as part of the shutwdown, and it would
cause the shutdown logic to be performed twice (the second
time the dll would not be there anymore).

Given that memcached loads all of the available engines during
startup and unloads them during shutdown it is cleaner to
let ewouldblock engine use the instance that is already loaded
than "fake load" of the engine and "silently drop" the engine
when it's done with it.

Change-Id: Idbb5ec09d487216fa29c254594acc5eef57a3055
Reviewed-on: http://review.couchbase.org/105066
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>