MB-33081: VBCBAdapter: Avoid double-locking of vbMap contents
VBCBAdapter currently takes a copy of the current vBucket ids from the
vbmap in it's constructor; then iterates this vector in its run()
method. This requires locking each vbucket twice - once to form the
list, and a second time to actually iterate the vBuckets in run().

Given that the state of a vBucket can change between taking a copy of
the VBids and actually iterating them, simplify by removing the copy
and just iterating directly on the vbMap inside run().

Change-Id: I336dddf033d9fef72b3372c066c81d6e2a0eee22
Reviewed-on: http://review.couchbase.org/105080
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>