Refactor: VBucketVisitor::visitBucket takes const shared_ptr&
Change VBucketVisitor::visitBucket() to take a const shared_ptr&,
instead of a non-const one. This makes the usage more explicit that
the visitBucket() function cannot reseat the shared_ptr; but it /may/
take a reference-count to it where necessary.

As part of updating usages, fix a couple of instances where:

a) A VBucketPtr (shared_ptr) was passed around uncessarily; a simple
   VBucket& reference was sufficient.

b) A VBucketPtr was unnecessarily copied (reference count taken), and
   the vbid was sufficient.

Change-Id: I91daf2cf0b396238bf2e7b627364d2c2fdded063
Reviewed-on: http://review.couchbase.org/105081
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>