MB-33153: Move topkeys unique ptr to reduce cache contention
The access of the topkeys unique ptr is write once, read often. The goal
is to share with other similar data, so that writes to the same cache
line containing this ptr are infrequent. Avoid sharing with other write
heavy objects such as the mutex, clients counter, or Timings.

Perf test (20/80 R/W):

Build 2384 - 1,831,406 ops/s (10 runs)
Build 2384 + this patch - 1,846,347 ops/s (10 runs)

Change-Id: I394f2302bc635a39cd1d9bd52ebf21f5ab99d437
Reviewed-on: http://review.couchbase.org/105169
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>