MB-33107: Fix intermittent failure in ObjectRegistryTest
The tests immediately before ObjectRegistryTest have an inconsistency
in the type used by loaderFun() - callers pass in a std::map, but
loaderFun() expects a std::set. This compiles (and apparently runs!)
as reinterpret_cast<> is used to convert between them.

However, when running on Windows wit jemalloc; this later results in
heap corruption manifesting in ObjectRegistryTest's SetUp method.

What happens is that when the std::map is deallocated the internal
nodes are not of the expected size, which causes cb_sized_free() and
ultimately je_sdallocx() with an incorrect size to free. This results
in the heap corruption.

Fix by using the correct type when calling loaderFun().

Change-Id: I03ea7623341eff8807c36938c0d94ee7ee6a3709
Reviewed-on: http://review.couchbase.org/105234
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>