Improve memory accounting of Checkpoint
Previously the getMemConsumption method returned only the size of memory
that was used by the items held in the queue, i.e. their key, metadata
and blob.  It did not include any of the overhead associated with
maintaining the checkpoint, for example the memory used to maintain the
keyIndex / metaKeyIndex.

The intention of the method is to return the total memory used by the
checkpoint and therefore it has been modified to also include the
overhead.

In addition the manual memory accounting has been reverted to the
approach used prior to commit f7eff02dc7f208cb2619e39a55d4827aacaa7ad6
(http://review.couchbase.org/#/c/103342/) which uses an approximate for
the overhead of the queue, as opposed to using the memory tracker.  This
is because in a future patch we will introduce the ability to expel
items from a checkpoint, and want the ability to release the underlying
memory associated with these items whilst the queueLock is NOT held.
This means we can end-up increasing memory usage as we add items to the
Checkpoint (in queueDirty), at the same time freeing memory. This means
we can't use the memory tracker functionality when doing manual memory
accounting. Note: manual memory accounting is not used in production.

The getMemoryOverhead method is modified to include the overhead of
the Checkpoint object.

The incrementMemConsumption and decrementMemConsumption methods have
also been removed.  This was for two reasons; (1) the names were
misleading (2) the methods are no longer invoked by CheckpointManager
and therefore the underlying member variable can be updated directly.

Change-Id: I42cbcf0736589faf2bc9d8beb46ecc16a5360a41
Reviewed-on: http://review.couchbase.org/105328
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>