MB-31811: Fix data race on collections persistedHighSeqno
In a test introduced in a future commit, we make many stats calls
to read the highSeqno of a collection. We don't actually care about
the persistedHighSeqno in these tests, but it is read by the stats calls
whilst things get flushed (this value gets written) in the bakground.
TSan identifies this as a data race which would cause a sporadic test
failure. Fix this by making the persistedHighSeqno AtomicMonotonic.

Change-Id: Iaaa6d172c57179f5523d8090fbdd49b3a7a3c48a
Reviewed-on: http://review.couchbase.org/105375
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>