MB-33226: OpenTracing prototype
This prototype allows you to hook into the OpenTracing
framework by having the client supply the parent span
context by using a new FrameInfo field in the header,
and the server adds the new spans by referencing that
as the parent.

It is configured through memcached.json in a new
section named "opentracing" which looks like:

  "opentracing": {
    "enabled": true,
    "module": "/opt/jaeger/lib/libjaegertracing.so.0.5.0",
    "config": "service_name: Couchbase"
  }

"enabled" is an on/off switch which may be toggled
at any time. If set to false the server silently
ignores any trace contexts provided in the request.

"module" contains the absolute path to a shared
library containing an implementation of the
OpenTracing API. This parameter may be changed at
any time, but once successfully loaded any change
in the parameter is ignored (the library will only
be loaded if enabled is also set to true).

"config" contains the configuration parameter
to pass on to the module when we try to create
the tracer. It parameter may be changed at any
time, but it is silently ignored if changed after
the module is successfully loaded.

See scripts/opentracing.sh for how to set the
parameters above.

Limitations:

Given that this is just a prototype it has a number
of limitations and shortcommings:

* No privilege support
  Clients should hold a privilege in order to enable
  tracing as it puts extra load on the server

* Integration with the tracer happens in the front
  end worker threads. It is not defined in the
  OpenTracing API if the calls may block or not.

* No relasionships with the stacked spans. All
  spans in the server is put directly under the
  client span.

* Limited information in each span. No effort
  was put into figuring out if we could add more
  context to the various spans.

Change-Id: I87506f36903f6dcd470c85abfb2c778c0f304ad6
Reviewed-on: http://review.couchbase.org/105383
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>