MB-33177: Extend chk_period for ep_testsuite checkpoint tests
 Address Sanitizer slows down the runtime of the tests and therefore
 it is possible that the test exceeds the default chk_period of 5s.
 This will cause a new checkpoint to be created and as the tests
 are not expecting this to happen they are prone to failure.

Most of the tests have already has their input parameters modified to
use a chk_period of 600.  This patch makes a similar change for the
remaining two tests that are vunerable to this issue.

- test_create_new_checkpoint
- test_checkpoint_persistence

The test_checkpoint_persistence also has its item_num_based_new_chk
input parameter explictly set to true, as it was missing the explict
assignment.

Change-Id: I38029ac3d77959c39d777a26c510fac4c853e831
Reviewed-on: http://review.couchbase.org/105417
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>