MB-33188 [SR]: Don't expire Pendings if Timeout is not set
As part of the Durability Requirements, the user can specify a Timeout
for a SyncWrite. The SyncWrite is aborted if Requirements are not
satisfied within the Timeout.

If Timeout is not specified, then the system assumes that the request
will never expire.

To encode this "no-timeout" scenario we use Timeout=0, which is also the
default if no other value is specified by the user.

Currently, Timeout=0 is processed as "expire immediately" rather than
"never expire".
With this patch we ensure that SyncWrites with Timeout=0 are never
expired.

Change-Id: I6337118b734daaf30c03a92313665b43f13c8c0c
Reviewed-on: http://review.couchbase.org/105472
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>