MB-33125: breakpad_test: Use fixed string for section delimiter
When gdb is invoked we run multiple commands, delimiting the output of
each command with an dummy string - the output of the 'show print
pretty' command.

On earlier versions of GDB this command prints:

    Pretty formatting of structures is off.

However on GDB 7.6.1 as shipped with CentOS 7.2 it prints a slightly
different output:

    Prettyprinting of structures is off.

This cases the output parsing to fail as it doesn't find the expected
delimiter:

    Traceback (most recent call last):
    File kv_engine/tests/breakpad_test.py, line 240, in <module>
        gdb_output[1], re.MULTILINE)
    IndexError: list index out of range

Fix by changing to a fixed delimiter string - a line of 80 hyphens.

Change-Id: Ifcf2063e806c66fb8ead130506132850d51066c2
Reviewed-on: http://review.couchbase.org/105576
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>