MB-33160 Decode error context in mc_bin_client
A response with a non-success status code can include an error context as the
returned value. Since the python3 conversion of mc_bin_client the response
value will be a bytes object, while the exception logic expects a string to
present to the user. Decode the error context before constructing the
exception.

Change-Id: I4fcaa1fbe68b4696c1c630363bece9483808080f
Reviewed-on: http://review.couchbase.org/105300
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>