[SR]: Add DiskDocKey class
Add DiskDocKey - a class wich represents a document key which has been
/ will be written to disk.

Objects of this type are created when we make a request to a KVStore
to read or write a document.

This is similar to DocKey (and related classes), except that may have
a 'DurabilityPrepare' namespace prefix prepended before the normal
collectionID if it refers to a document in the separate prepare
namespace.  As such, it *cannot* be implicitly converted to DocKey -
the user must consider if they explicitly want to ignore the
DurabilityPrepare namespace.

This patch adds DiskDocKey and unit tests, but doesn't yet make use of
it in the rest of ep-engine.

Change-Id: Iac0a7450f7c6e5805fc6a3e5445be8f6e92c0dd9
Reviewed-on: http://review.couchbase.org/105717
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>