MB-31664: Reapply - Switch over to not using backfill queue
The long-term plan is to remove the backfill queue.  However, to gain
confidence in the change prior to its removal we are disabling via the
use of a switch.

This switch over was originally done in
commit 5968b4beaae496e080cbf174259ef01a7dabdabc
(see http://review.couchbase.org/#/c/101275/) however it had to be
reverted in commit 3386b902bd90c4c113a3f732cab0bf64591c08aa
(see http://review.couchbase.org/101684) because it was discovered that
for a backfill, the checkpoint in the checkpointManager could become so
large that we end up hitting an OOM condition.

With the introducion of the ability to expel items from a checkpoints
where all cursor have past, this should no longer occur.  Therefore it
should be safe to disable the use of the backfill queue.

If after a period of time no issues are seen, the backfill queue and
associated switch will be removed.

Change-Id: I188cfbb48edbd4cf7279fa32f7933722e2d9577a
Reviewed-on: http://review.couchbase.org/105643
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>