MB-32687: Backport signalIfIdle refactor
signalIfIdle used to update the event mask in libevent to add
a write event to make the socket available. Libevent would then
pick up that socket next time it went down to the OS to do a poll
on the sockets. The problem is that when we try to update the
event mask for a socket libevent will try to acquire the mutex
used by the event base in order to do that. If the code was
called from another worker thread it would hold it's own lock
on the event base (and we're subject for a deadlock).

The new implementation of signalIfIdle push the connection
onto a list of connections to signal for the worker thread,
before it notifies the worker thread by writing to the
notification pipe for the worker thread. The callback for
the worker thread swaps out the list of connections, and
runs the state machine for each of the connections in there.

Change-Id: Ic0ba5de83b8f924a8b7457c65d0e91fbf7927f7d
Reviewed-on: http://review.couchbase.org/105714
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>