[SR]: Use DiskDocKey for writing to disk
Update IORequest to use a DiskDocKey for the to-be-written key,
removing the need for the separate 'pending' flag.

In turn, StoredDocKey no longer needs to allow creation with the
'pending' flag - DiskDocKey handles this use-case now - so revert
StoredDocKey back to previous behaviour. CollectionID ctor now
considers DurabilityPrepare a reserved namespace and throws if an
attempt is made to create a CollectionID "normally" with that
namespace.

Change-Id: Ibf45086267f772cf9e6b15345cf78a39b54fcdee
Reviewed-on: http://review.couchbase.org/105727
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>