[SR]: Use DiskDocKey for compaction hooks
Use DiskDocKey for compaction-related callbacks for (a) deleting
expired items and (b) dropping keys which are from deleted
collections.

TTLs are not applied to prepared SyncWrites, so the work here is
mostly to re-wire for the new type.

Collection deletion /does/ however affect prepared SyncWrites, as when
a collection is dropped we must remove all items (including prepared)
which exist for it.

Change-Id: Iee021617dbbc29efc4a35d340c478e8a02aabb64
Reviewed-on: http://review.couchbase.org/105729
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>