MB-33244 [SR]: Don't de-dupe mutations with prepared SyncWrite in Checkpoint
When a Prepared SyncWrite is enqueued into a Checkpoint it can
incorrectly de-duplicate an existing CommitedViaMutation item - i.e. a
normal mutation.

This is incorrect, as the prepared SyncWrite may eventually be
aborted, and such it's result should be discarded. We therefore need
to keep both in the Checkpoint to send over DCP & persist to disk.

Change-Id: I1cdb4b7407fbccb94ca4df5a8e8d922ead899256
Reviewed-on: http://review.couchbase.org/105730
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>