Warmup: Remove dead code in keyDumpforShard
Warmup::keyDumpforShard() checks if all shards set their status to
true in shardKeyDumpStatus, before reporting success for this
stage. However this is a tautology - shardKeyDumpStatus is always set
to true.

This was not always the case - prior to 797c9e0d5 (MB-14191: Implement
warmup using ForestKVStore - included in v4.5.0) the logic was
different and if the storage engine didn't support keyDump then
shardKeyDumpStatus could be false.

Simplify the logic by constant-propogating into shardKeyDumpStatus.

Change-Id: Iffabbad4a0ced7ed5010a0225d8e72e25319d34a
Reviewed-on: http://review.couchbase.org/105870
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>