MB-33226: Push trace from a dedicated thread
To avoid OpenTracing to potentially block the front-end
threads while the OpenTracing module performs the operations
we'll push the context to a "list" and process those entries
in a dedicated thread.

NOTE: The code assumes that only a subset of the commands
requests OpenTracing otherwise we'll probably get lock
contention (and should migrate over to a "per thread"
list).

Change-Id: Ib82e07bbd77a3e1fea8e4368839319ce81bb51e9
Reviewed-on: http://review.couchbase.org/106116
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>