MB-32763: test_disk_gt_ram_golden: remove mem_used check
Testcase `test_disk_gt_ram_golden` checks various stats during a basic
DGM scenario; however it makes assumptions about mem_used which are
incorrect. Specifically, it assumes that mem_used will only decrease
when checking before / after an item is evicted; and similiar that it
will be restored to the original value after bgfetch.

While this is reasonable if ep-engine was entirely synchronous, this
is not the case as background tasks can also allocate & free
memory. As such, we see intermittent test failures of this test where the
mem_used doesn't change as expected.

Given there are other synchronous tests which already check memory
tracking, remove the mem_used checks from this testcase.

Change-Id: I9d57efc056564d79dfb79ef40818eb8cd20933ef
Reviewed-on: http://review.couchbase.org/106003
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>