MB-33298 [SR]: Ensure seqno-invariant in DurabilityMonitor
At Vucket::queueDirty we do 2 things:
1) we queue an item into the BackfillQueue/CheckpointManager
2) and we add durable items into the DurabilityMonitor

While every structure ensures internal synchronization, it could happen
that we queue e.g. seqno:1 and seqno:2 into the CM, but seqno:2 first
and then seqno:1 into the DM for tracking, which is wrong.
Given that VBucket::queueDirty requires only a HashBucketLock, that
issue may arise when multiple threads issue durable-writes that fall
into different HashTable partitions.

This patch ensures that:
a) a proper exception is thrown if any actor tries to break
  seqno-invariant when queueing a new item into the DM
b) at VBucket::queueDirty we treat enqueueing in Bakcfill/CM and DM as
  an atomic operation. That ensures that items are queued into all
  structures with the same seqno-order.

Change-Id: Ibe589925d676ed90940d1dae3a3b3f3fdd9c9f8a
Reviewed-on: http://review.couchbase.org/106065
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>