MB-33372: Disable Checkpoint Expel Items
The master-post-commit test (make simple test) is failing due to
waiting for the replica item count to match in xdcrnewbasetests.
This is believed to be due to the recent introduction of expelling
checkpoint items. See http://review.couchbase.org/#/c/104945/
with Change-Id: I846aa70541d1f6f3952cd53334efe6698d683577

Therefore temporarily disable the checkpoint expelling feature until the
cause of the issue is resolved.

Change-Id: If12c4b98e570cd2c9db8366416d894afdbf78b39
Reviewed-on: http://review.couchbase.org/106267
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>