MB-33739: Use reference in ConnMap (and DcpConnMap) for vbConns
This should reduce cache contention on the shared_ptr as we
no longer need to acquire it (write to the ownership counter)
when we create a new seqno. Given that we were writing to the
same producer for every VB (in a 2 node test) the contention
was extreme. The lifetime of the object is still managed by the
shared_ptr, we've only replaced weak_ptrs with references, and on
shutdown of Consumers or Producers they remove themselves from
vbConns.

Cache contention on the shared_ptr to the DcpProducer dropped from ~5%
to ~0.5%, which is now coming from ActiveStream.

Perf stats (ops/s):
Triton Cluster - Pillowfight, 20/80 R/W, 256B binary items. 2 Nodes.

Current master - ~1,965,984 ops/s (6 run average)
With this patch - 1,992,394 ops/s (10 run average)

Change-Id: I2fad36416b61359794462025424f9694a5bab352
Reviewed-on: http://review.couchbase.org/106345
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>