MB-33296, MB-33172 [SR]: DM enforces durability-impossible
This patch fixes the definition of Majority in DurabilityMonitor:
  majority = chain_size / 2 + 1
always, even in the case where the chain contains Undefined nodes.

With this patch the DM checks if 'durability-impossible' every time a
new write is added for tracking.

Note that a proper error code is returned by the DM but not yet
processed by the caller (VBucket). That means, the clients will not
receive an error code yet. That is added in a dedicated patch.

Note that this patch implicitly fixes MB-33172. From now on, the
DurabilityMonitor returns durability-impossible if the user issues a
SyncWrite but the replication-chain is not set (rather than throwing an
expection).

Change-Id: I265dcb56ca9a0432e941daf6e49bbad8733857ac
Reviewed-on: http://review.couchbase.org/106351
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>