MB-33372: Fix bug where expelling item needed by cursor
If we have two cursors; one pointing to a mutation and the other
pointing to a meta-data item immediately preceeding the mutation, as
shown in the following example for dcpCursor1 and dcpCursor2.

     1000 - dummy item
     1001 - checkpoint start  <<<<<<<< dcpCursor1
     1001 - 1st item  <<<<<<< dcpCursor2
     1002 - 2nd item  <<<<<<< persistenceCursor

Previously if we set dcpCursor2 as the initial lowest cursor we
would not replace with dcpCursor1 because it has the same seqno.

This patch ensures that in such a case where there are two cursors both
with the same seqno, but with one pointing to the meta-data item, and
the other pointing to the mutation, the lowest cursor is updated to
refer to the cursor pointing to the meta-data item.

Change-Id: I780ee77cc12c67c0c7632a689f59b576f1418c21
Reviewed-on: http://review.couchbase.org/106363
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>