Remove unneeded IORequest virtual dtor
While IORequest is subclassed by the different kvstores, we never
manipulate different class types - e.g. we don't put a mix of
CouchRequest / RocksRequest in the same container and need to destruct
each differently.

As such, remove the virtual keyword from the IORequest and
CouchRequest dtor. This doesn't yet make the objects smaller, but does
allow for future optimizations in size.

Change-Id: I9be72fc1c7c32574fb5c1daac98c0d4e635e03d2
Reviewed-on: http://review.couchbase.org/106376
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>