MB-33423: Temporarily replace exception with abort
CheckpointIterator contains exception code to cover the unexpected
case of attempting to access from the end() position.  As this is
invalid an exception is thrown.

This exception was triggered during system test however the undelying
cause has yet to be established.  Therefore adding detailed logging and
temporarily replacing the throwing of the exception with an std::abort
so can debug if the issue occurs again during testing.

It is intended that std:abort will be reverted back to throwing an
exception prior to the release of Mad-Hatter.

Change-Id: I6eb0f00e6c900a11946d414ef04e86add97e0019
Reviewed-on: http://review.couchbase.org/106703
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>