MB-33495: Only pause AccessScanner for 0s when duration cap hit
The AccessScanner uses a subclass of CappedDurationVBucketVisitor to
visit vBuckets and construct an access log; yielding to other tasks if
visiting takes longer than 25ms.

CappedDurationVBucketVisitor was introduced in 8ea82f923, however in
the case of AccessAcanner the yield was incorrectly set to the access
log interval (default 1 day!), instead of essentially immediately
(0s).

*If* the ItemAccessVisitor tasks visits all vBuckets in under 25ms
then it still works correctly, however if not (which is likely on any
non-toy dataset) then the ItemAccessVisitor will end up sleeping for a
default of 1 day after every 25ms of work, most likely resulting in
the access.log failing to be generated for many, many days.

Fix by putting the sleep back to the correct 0s (i.e. just yield for
other higher-priority tasks to run if waiting).

Change-Id: Ic0d143948342e61ba3c20a8464a6d3c0dcb74639
Reviewed-on: http://review.couchbase.org/106714
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>