MB-33463: Add new magma storage engine configuration parameters
Add the configuration parameters necessary to configure magma.

Most of the configuration parameters are pass thru except for
MaxWriteCacheSize. MaxWriteCacheSize is the buffer that holds the items
prior to them being flushed to the sstables. The items are first flushed
to the WAL for persistence and the item is held in a shard level
skiplist until its flushed to an sstable.

WriteCacheSize = BucketQuota / # shards * Mem Quota Ratio
AND
WriteCacheSize >= 8MB <= 128MB.

Additional http://review.couchbase.org/107094 checkin done to magma to
link EP engine config parameters document. Leaving as google sheet for
now while magma remains in a state of flux and once we have completed
the necessary changes, testing and review comments, we can fold the
sheet into the magma README.md file.

Change-Id: Icdfa680ae4539c1737124218add903a4c048ec95
Reviewed-on: http://review.couchbase.org/106745
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>