Fix misleading placement of '#pragma pack(1)'
As identified via http://review.couchbase.org/#/c/106745/, '#pragma
pack(1)' must be placed before the definition of structs & classes -
placing it before specific members of a class does not have any
effect.

A number of uses inside kv-engine - specifically in the KVStore
backends was incorrect - the pragma occurred in the middle of the
class. However, in all cases the misplaced pragma was either benign
because either:

- The natural alignment of the object was already packed
  (couch-kvstore-metadata.h) - note there are static_asserts()
  checking the size; or

- The struct(s) didn't _need_ to be packed (rocksdb-kvstore.cc /
  magma-kvstore.cc) - because we don't (yet) need backwards compatible
  on-disk layout.

Nethertheless, fix all incorrect usages of #pragma pack to not mislead
anyone. (magma-kvstore.cc is undergoing parallel changes so not
modifying those in this patch).  Additionally reduce the size of the
type used for sequence numbers in rockskv::MetaData as we only need
48bits.

Change-Id: I6f9d53a5668066210ad0f80250d14c2bfaea2820
Reviewed-on: http://review.couchbase.org/106846
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>